requirements.txt
streamlit
pandas
google-genai
scikit-learn